split PlainTextFormatterTests into smaller files
